style updates and refactor
