+++ client/src/app/Authorized/PageContent/Transactions/TransactionCards.tsx/TransactionCard/TransactionCard.module.css
@@ -7,45 +7,3 @@
     outline: 1px solid var(--mantine-primary-color-filled);
   }
 }
-
-.container {
-  width: 100%;
-  gap: 0.5rem;
-  align-items: center;
-  flex-direction: row;
-
-  @container (max-width: 700px) {
-    flex-direction: column;
-  }
-}
-
-.subcontainer {
-  gap: 0.5rem;
-  align-items: center;
-
-  @container (max-width: 700px) {
-    width: 100%;
-  }
-}
-
-.dateContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-}
-
-.merchantContainer {
-  flex: 1 1 auto;
-}
-
-.categoryContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-}
-
-.amountContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: start;
-}
