+++ client/src/app/Authorized/PageContent/Transactions/TransactionCards.tsx/TransactionCards.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import TransactionCard from "./TransactionCard/TransactionCard";
 import {
   defaultTransactionCategories,
   Filters,
@@ -16,6 +15,7 @@ import { AuthContext } from "~/components/AuthProvider/AuthProvider";
 import { useQuery } from "@tanstack/react-query";
 import { AxiosResponse } from "axios";
 import { ICategoryResponse } from "~/models/category";
+import TransactionCardContent from "./TransactionCard/TransactionCard";
 
 interface TransactionCardsProps {
   filters: Filters;
@@ -92,7 +92,7 @@ const TransactionCards = (props: TransactionCardsProps): React.ReactNode => {
                 (page - 1) * itemsPerPage + itemsPerPage
               )
               .map((transaction) => (
-                <TransactionCard
+                <TransactionCardContent
                   key={transaction.id}
                   transaction={transaction}
                   categories={transactionCategoriesWithCustom}
