+++ client/src/components/EditableCategoryCell/EditableCategoryCell.tsx
